chore(test): use output interface in dataset import command, move test to integration
